+++ app/src/main/java/uz/gita/jaxongir/sellmanageradmin/data/requests/ProductRequest.kt
@@ -4,7 +4,7 @@ data class ProductRequest(
     val id : String,
     val name : String,
     val count : Int,
-    val initialPrice : Int,
+    val initialPrice : Double,
     val isValid : Boolean,
     val comment : String
 )
